// src/redux/action-types.ts
export enum ActionTypes {
    ADD_MOVIE = 'ADD_MOVIE',
    DELETE_MOVIE = 'DELETE_MOVIE',
    UPDATE_MOVIE = 'UPDATE_MOVIE',
    LOAD_MOVIES = 'LOAD_MOVIES',
    SET_MOVIES = 'SET_MOVIES',
    GET_MOVIE = 'GET_MOVIE',
    SET_MOVIE = 'SET_MOVIE',
    SET_ALERT = 'ADD_ALERT',
    SET_LOADING = 'SET-LOADING',
    SET_MOVIE_NOT_FOUND = 'SET_MOVIE_NOT_FOUND',
    SET_FILTER = 'SET_FILTER',
    SET_SORT_ORDER = 'SET_SORT_ORDER',
    SET_SEARCH = 'SET_SEARCH',
    SET_SORT_BY = 'SET_SORT_BY',
    SET_QUERY_PARAMS = 'SET_QUERY_PARAMS'
}